split simple and non simple seed tests
